components/dags/pyproject.toml
[project]
name = "pinta-dags"
version = "0.0.0"
description = "DAG component of the elevation production system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pinta-processing",
]
[project.optional-dependencies]
airflow = [
  "apache-airflow[postgres,standard]==3.1.8",
]

[build-system]
requires = ["uv_build>=0.10.9,<0.12.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "pinta_dags"
include = ["py.typed"]

[tool.uv.sources]
pinta-processing = { workspace = true }

[tool.uv]
default-groups = ["dev", "lint"]

[dependency-groups]
dev = [
  # Airflow as dev dependency to avoid pinning the version
  # of published dag package to specific airflow version
  # (airflow can be updated without releasing a new version of the dags)

  "pytest-dotenv",
  "pytest-mock",
  "pytest-order",
  "pytest-rerunfailures",
  "pytest-subtests",
  "pytest-timeout",
  "pytest-xdist",
  "coverage-enable-subprocess",
]
lint = [
  "prek",
  "mypy",
  "ruff",
]

[tool.pytest.ini_options]
testpaths = [
  "test",
]
addopts = "-n 0 --import-mode=importlib"

[tool.coverage.paths]

[tool.coverage.run]
source_pkgs = ["pinta_dags"]
parallel = true
relative_files = false
# TODO: set branch=true and remove core="sysmon" when python 3.14
# is used (coverage will use sysmon by default)
branch = false
core = "sysmon"

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "if TYPE_CHECKING:",
  "raise NotImplementedError",
]

# extend required for isort action to pick up correct config
[tool.ruff]
extend = "../../pyproject.toml"

[tool.semantic_release]
commit_parser = "conventional-monorepo"
major_on_zero = false
allow_zero_version = true
commit_message = """\
chore(release): pinta-dags@{version}

Automatically generated by python-semantic-release
"""
tag_format = "pinta-dags-v{version}"
version_toml = ["pyproject.toml:project.version"]
build_command = """
    python -m pip install -e '.[build]'
    rm -f dist
    mkdir -p dist
    uv lock --upgrade-package "$PACKAGE_NAME"
    git add ../../uv.lock
    uv build
    mv ../../dist/pinta_dags* dist/
"""

[tool.semantic_release.commit_parser_options]
path_filters = ["."]
scope_prefix = "dags"

[tool.semantic_release.changelog]
# Recommended patterns for conventional commits parser that is scope aware
exclude_commit_patterns = [
  '''chore(?:\([^)]*?\))?: .+''',
  '''ci(?:\([^)]*?\))?: .+''',
  '''refactor(?:\([^)]*?\))?: .+''',
  '''style(?:\([^)]*?\))?: .+''',
  '''test(?:\([^)]*?\))?: .+''',
  '''build\((?!deps\): .+)''',
  '''Initial [Cc]ommit.*''',
]

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease_token = "dev"
